+++ pkg/datastream/stream_activation_test.go
@@ -61,7 +61,7 @@ func TestDs_ActivateStream(t *testing.T) {
 				StreamStatus: StreamStatusActivating,
 				DeliveryConfiguration: DeliveryConfiguration{
 					Format:    FormatTypeJson,
-					Frequency: Frequency{IntervalInSeconds: IntervalInSeconds30},
+					Frequency: Frequency{IntervalInSeconds: 30},
 				},
 				Destination: Destination{
 					CompressLogs:    true,
@@ -160,7 +160,7 @@ func TestDs_ActivateStream(t *testing.T) {
 					Delimiter: DelimiterTypePtr(DelimiterTypeSpace),
 					Format:    FormatTypeStructured,
 					Frequency: Frequency{
-						IntervalInSeconds: IntervalInSeconds30,
+						IntervalInSeconds: 30,
 					},
 					UploadFilePrefix: "ak",
 					UploadFileSuffix: "ds",
@@ -326,7 +326,7 @@ func TestDs_DeactivateStream(t *testing.T) {
 				StreamStatus: StreamStatusDeactivating,
 				DeliveryConfiguration: DeliveryConfiguration{
 					Format:    FormatTypeJson,
-					Frequency: Frequency{IntervalInSeconds: IntervalInSeconds30},
+					Frequency: Frequency{IntervalInSeconds: 30},
 				},
 				Destination: Destination{
 					CompressLogs:    true,
@@ -425,7 +425,7 @@ func TestDs_DeactivateStream(t *testing.T) {
 					Delimiter: DelimiterTypePtr(DelimiterTypeSpace),
 					Format:    FormatTypeStructured,
 					Frequency: Frequency{
-						IntervalInSeconds: IntervalInSeconds30,
+						IntervalInSeconds: 30,
 					},
 					UploadFilePrefix: "ak",
 					UploadFileSuffix: "ds",